Move git to filesystems
